+++ de.metas.business/src/main/java/de/metas/costing/methods/LastPOCostingMethodHandler.java
@@ -83,7 +83,7 @@ public class LastPOCostingMethodHandler extends CostingMethodHandlerTemplate
 		{
 			if (qty.signum() != 0)
 			{
-				final CostAmount price = amt.divide(qty, currentCosts.getPrecision(), RoundingMode.HALF_UP);
+				final CostAmount price = amt.divide(qty, currentCosts.getPrecision().toInt(), RoundingMode.HALF_UP);
 				currentCosts.setCostPrice(CostPrice.ownCostPrice(price));
 			}
 			else
